requirements.txt
transformers
Rouge
nltk
gensim
bert_score
tensorflow
bleurt
unbabel-comet
tensorflow
pytorch
numpy